Modifica Ops para operar con los cambios en el backend.

+++ src/main/java/sic/vista/swing/DetalleNotaDebitoGUI.java
@@ -516,24 +516,29 @@ public class DetalleNotaDebitoGUI extends JDialog {
                     + "/pago/" + pago.getId_Pago(), notaDebito, NotaDebito.class);
             if (notaDebito != null) {
                 notaDebitoCreada = true;
-                if (Desktop.isDesktopSupported()) {
-                    try {
-                        byte[] reporte = RestClient.getRestTemplate()
-                                .getForObject("/notas/" + notaDebito.getIdNota() + "/reporte",
-                                        byte[].class);
-                        File f = new File(System.getProperty("user.home") + "/NotaDebito.pdf");
-                        Files.write(f.toPath(), reporte);
-                        Desktop.getDesktop().open(f);
-                    } catch (IOException ex) {
-                        LOGGER.error(ex.getMessage());
+                int reply = JOptionPane.showConfirmDialog(this,
+                        ResourceBundle.getBundle("Mensajes").getString("mensaje_reporte"),
+                        "Aviso", JOptionPane.YES_NO_OPTION);
+                if (reply == JOptionPane.YES_OPTION) {
+                    if (Desktop.isDesktopSupported()) {
+                        try {
+                            byte[] reporte = RestClient.getRestTemplate()
+                                    .getForObject("/notas/" + notaDebito.getIdNota() + "/reporte",
+                                            byte[].class);
+                            File f = new File(System.getProperty("user.home") + "/NotaDebito.pdf");
+                            Files.write(f.toPath(), reporte);
+                            Desktop.getDesktop().open(f);
+                        } catch (IOException ex) {
+                            LOGGER.error(ex.getMessage());
+                            JOptionPane.showMessageDialog(this,
+                                    ResourceBundle.getBundle("Mensajes").getString("mensaje_error_IOException"),
+                                    "Error", JOptionPane.ERROR_MESSAGE);
+                        }
+                    } else {
                         JOptionPane.showMessageDialog(this,
-                                ResourceBundle.getBundle("Mensajes").getString("mensaje_error_IOException"),
+                                ResourceBundle.getBundle("Mensajes").getString("mensaje_error_plataforma_no_soportada"),
                                 "Error", JOptionPane.ERROR_MESSAGE);
                     }
-                } else {
-                    JOptionPane.showMessageDialog(this,
-                            ResourceBundle.getBundle("Mensajes").getString("mensaje_error_plataforma_no_soportada"),
-                            "Error", JOptionPane.ERROR_MESSAGE);
                 }
                 this.dispose();
             }
